server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
pydantic[email]==2.5.3
python-dotenv==1.0.0

# Langchain (your AI dependencies)
langchain==0.1.4
langchain-community==0.0.16
langchain-mistralai==0.0.4
# Or whatever AI provider you're using
